[UF-364] uberfire-security: add support for WildFly10/EAP7
 * both WildFly8/EAP6 and WildFly10/EAP7 are now supported.
   They need to be in separate modules, because the WildFly
   libs in use are not fully backwards compatible (API wise)

 * the -wildfly-commons module contains shared code to avoid
   duplication
