+++ uberfire-security/uberfire-security-management/uberfire-security-management-wildfly/uberfire-security-management-wildfly-commons/src/main/java/org/uberfire/ext/security/management/wildfly/properties/WildflyPropertiesFileLoader.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+package org.uberfire.ext.security.management.wildfly.properties;
+
+import java.io.IOException;
+import java.util.Properties;
+
+/**
+ * Simple abstraction over the WildFly's PropertiesFileLoader in order to support different implementations for
+ * different WildFly versions
+ *
+ * Specific WildFly/EAP providers should implement this interface.
+ */
+public interface WildflyPropertiesFileLoader {
+
+    Properties getProperties() throws IOException;
+
+    void persistProperties() throws IOException;
+
+    void stop();
+}
