+++ uberfire-security/uberfire-security-management/uberfire-security-management-wildfly/uberfire-security-management-wildfly10/src/test/resources/org/uberfire/ext/security/management/wildfly10/application-roles.properties
