+++ uberfire-security/uberfire-security-management/uberfire-security-management-wildfly/uberfire-security-management-wildfly8/src/test/resources/org/uberfire/ext/security/management/wildfly8/application-users.properties
@@ -1,0 +1,40 @@
+#
+# Copyright 2016 Red Hat, Inc. and/or its affiliates.
+#  
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#  
+#    http://www.apache.org/licenses/LICENSE-2.0
+#  
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+#
+# Properties declaration of users for the realm 'ApplicationRealm' which is the default realm
+# for application services on a new installation.
+#
+# This includes the following protocols: remote ejb, remote jndi, web, remote jms
+#
+# Users can be added to this properties file at any time, updates after the server has started
+# will be automatically detected.
+#
+# The format of this realm is as follows: -
+#
+# A utility script is provided which can be executed from the bin folder to add the users: -
+# - Linux
+#  bin/add-user.sh
+#
+# - Windows
+#  bin\add-user.bat
+# The following illustrates how an admin user could be defined, this
+# is for illustration only and does not correspond to a usable password.
+#
+admin=0cd938691c1a7cef8722398281eecdbe
+user1=9ee49fd51e158b42ba273737eea4a813
+user2=204764b333e04f93a099c4cf1dc1e64b
+user3=8f0ef3372fbf29e6317f3e397a532b8f
